Put the plug-ins into separate credits + initialized crates for the logic of the server and pipelines
