+++ backend/bin/http_server.rs
@@ -1,0 +1,5 @@
+use http_server::test;
+
+fn main() {
+    test();
+}
