+++ backend/bin/pipeline_consumer.rs
@@ -1,0 +1,6 @@
+use pipeline_consumer::run_consumer;
+
+#[tokio::main]
+async fn main() {
+    run_consumer().await
+}
