+++ backend/core/Cargo.toml
@@ -1,0 +1,6 @@
+[package]
+name = "core"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
