+++ backend/extensions/facial_detecting/src/lib.rs
@@ -1,4 +1,5 @@
 pub mod config;
 pub mod models;
-pub mod pipeline;
+pub mod post_processing;
+pub mod predictor;
 pub mod utils;
