+++ backend/extensions/facial_recognizer/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "facial_recognizer"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+ndarray.workspace = true
+nalgebra.workspace = true
+image.workspace = true
+ort.workspace = true
+serde.workspace = true
+facial_detecting.workspace = true
+dotenv_config.workspace = true
