+++ backend/extensions/facial_recognizer/src/lib.rs
@@ -1,4 +1,4 @@
 pub mod config;
 pub mod models;
-pub mod pipeline;
+pub mod predictor;
 pub mod utils;
