+++ backend/extensions/facial_recognizer/src/utils.rs
