+++ backend/extensions/image_embedding/src/lib.rs
@@ -1,0 +1,3 @@
+pub mod config;
+pub mod models;
+pub mod predictor;
