+++ backend/extensions/image_embedding/src/predictor.rs
