+++ backend/extensions/text_embedding/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "text_embedding"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+ndarray.workspace = true
+nalgebra.workspace = true
+image.workspace = true
+ort.workspace = true
+serde.workspace = true
+facial_detecting.workspace = true
+dotenv_config.workspace = true
+itertools.workspace = true
+tokenizers = { version = "0.19", features = ["hf-hub", "http"] }
