+++ backend/extensions/text_embedding/src/predictor.rs
