+++ backend/http_server/Cargo.toml
@@ -1,0 +1,6 @@
+[package]
+name = "http_server"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
