--- backend/pipeline/src/ml/facial_processing/detection/mod.rs
@@ -1,2 +1,0 @@
-pub mod post_processing;
-pub mod predictor;
