--- backend/pipeline/src/ml/facial_processing/mod.rs
@@ -1,8 +1,0 @@
-mod detection;
-mod recognition;
-mod swap;
-mod transforms;
-
-pub use detection::predictor::FaceDetector;
-pub use recognition::predictor::FaceRecognizer;
-pub use transforms::umeyama;
