--- backend/pipeline/src/ml/facial_processing/recognition/mod.rs
@@ -1,1 +1,0 @@
-pub mod predictor;
