--- backend/pipeline/src/ml/mod.rs
@@ -1,2 +1,0 @@
-pub mod facial_processing;
-pub mod search;
