--- backend/pipeline/src/ml/search/mod.rs
@@ -1,5 +1,0 @@
-mod textual;
-mod visual;
-
-pub use textual::predictor::ImageTextualize;
-pub use visual::predictor::ImageVisualize;
