--- backend/pipeline/src/ml/search/textual/mod.rs
@@ -1,1 +1,0 @@
-pub mod predictor;
