--- backend/pipeline/src/ml/search/visual/mod.rs
@@ -1,1 +1,0 @@
-pub mod predictor;
