--- backend/pipeline/src/services/mod.rs
@@ -1,3 +1,0 @@
-pub mod facial_processing;
-pub mod pipeline;
-pub mod search;
