+++ backend/pipeline_consumer/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "pipeline_consumer"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+pipeline.workspace = true
+rdkafka.workspace = true
+uuid.workspace = true
+tokio.workspace = true
+serde_json.workspace = true
+dotenv_config.workspace = true
+dotenvy.workspace = true
